Moving the calculations to a shared package
