+++ app/src/main/java/com/brewthings/app/ui/screens/pill/insights/InsightsConverter.kt
@@ -3,6 +3,8 @@ package com.brewthings.app.ui.screens.pill.insights
 import com.brewthings.app.data.domain.Insight
 import com.brewthings.app.data.model.RaptPillData
 import com.brewthings.app.data.model.RaptPillInsights
+import com.brewthings.app.util.calculateABV
+import com.brewthings.app.util.calculateVelocity
 import com.brewthings.app.util.datetime.TimeRange
 import com.brewthings.app.util.datetime.daysBetweenIgnoringTime
 import kotlin.math.abs
@@ -98,18 +100,4 @@ private fun RaptPillData.toInsights(
         isOG = pillData.isOG,
         isFG = pillData.isFG,
     )
-}
-
-private fun calculateABV(og: Float, fg: Float): Float {
-    if (og <= 1.0 || fg <= 1.0) return 0f
-    return (og - fg) * 131.25f
-}
-
-private fun calculateVelocity(ogData: RaptPillData, fgData: RaptPillData): Float? {
-    val gravityDrop = fgData.gravity - ogData.gravity
-    val timeDifference = daysBetweenIgnoringTime(fgData.timestamp, ogData.timestamp).toFloat()
-    val velocity = gravityDrop / timeDifference
-    return if (velocity.isInfinite() || velocity.isNaN()) {
-        null
-    } else velocity
 }
